[GEOS-11433] Upgrade maven-pmd-plugin to 3.23.0 and PMD 7.2.x

+++ src/extension/sldService/src/main/java/org/geoserver/sldservice/utils/classifier/impl/CustomColorRamp.java
@@ -131,9 +131,9 @@ public class CustomColorRamp implements ColorRamp {
     private void interpolate(
             List<Color> result, Color start, Color end, int samples, boolean offset) {
         if (offset) {
-            double sRed = ((double) end.getRed() - start.getRed()) / (double) (samples);
-            double sGreen = ((double) end.getGreen() - start.getGreen()) / (double) (samples);
-            double sBlue = ((double) end.getBlue() - start.getBlue()) / (double) (samples);
+            double sRed = ((double) end.getRed() - start.getRed()) / samples;
+            double sGreen = ((double) end.getGreen() - start.getGreen()) / samples;
+            double sBlue = ((double) end.getBlue() - start.getBlue()) / samples;
 
             start =
                     new Color(
@@ -148,9 +148,9 @@ public class CustomColorRamp implements ColorRamp {
         int red;
         int green;
         int blue;
-        double sRed = ((double) end.getRed() - start.getRed()) / (double) (samples - 1);
-        double sGreen = ((double) end.getGreen() - start.getGreen()) / (double) (samples - 1);
-        double sBlue = ((double) end.getBlue() - start.getBlue()) / (double) (samples - 1);
+        double sRed = ((double) end.getRed() - start.getRed()) / (samples - 1);
+        double sGreen = ((double) end.getGreen() - start.getGreen()) / (samples - 1);
+        double sBlue = ((double) end.getBlue() - start.getBlue()) / (samples - 1);
         for (int i = 0; i < samples; i++) {
             red = (int) (sRed * i + start.getRed());
             green = (int) (sGreen * i + start.getGreen());
